Release 2025.8.1
# Release 2025.8.1

## 📢 Important Notice
My WindShares ends on September 1st, and after that I will no longer be able to test this custom component.

I’m looking for someone in the community who can help with testing and releasing future versions. If no one is able to take over testing, this integration will unfortunately be archived on GitHub and removed from HACS.

## Added
- Nothing

## Updated
- Boto3 1.38.7 -> 1.40.11

## Removed
- Nothing